split the subcommands into their own files.
