+++ src/lib.rs
@@ -8,7 +8,10 @@ pub mod command;
 pub mod container;
 pub mod create;
 pub mod dbus;
+pub mod delete;
 pub mod info;
+pub mod kill;
+pub mod list;
 pub mod logger;
 pub mod namespaces;
 pub mod notify_socket;
@@ -18,6 +21,7 @@ pub mod rootfs;
 pub mod rootless;
 pub mod signal;
 pub mod start;
+pub mod state;
 pub mod stdio;
 pub mod tty;
 pub mod utils;
